requirements.txt
streamlit
firebase_admin
pinecone-client
langchain
streamlit_option_menu
python-dotenv
PyPDF2
langchain_openai
langchain-community
sklearn
translate
langdetect